LOTS of stuff

+++ Files/Events/messageEvents/messageCreate/willis.js
@@ -1,21 +1,11 @@
 module.exports = {
   async execute(msg) {
     if (!msg.channel || msg.channel.type === 1) return;
-    if (msg.channel.id !== '805839305377447936' || !msg.author) return;
-    if (msg.attachments.size === 0) {
-      if (msg.member) {
-        if (
-          !msg.member.roles.cache.has('278332463141355520') &&
-          !msg.member.roles.cache.has('293928278845030410') &&
-          !msg.member.roles.cache.has('768540224615612437')
-        ) {
-          msg.react('✅').catch(() => {});
-          msg.react('❌').catch(() => {});
-        }
-      }
-    } else if (
+    if (msg.channel.id !== '979811225212956722' || !msg.author) return;
+    if (
       !msg.member.roles.cache.has('293928278845030410') &&
-      !msg.member.roles.cache.has('278332463141355520')
+      !msg.member.roles.cache.has('278332463141355520') &&
+      msg.attachments.size === 0
     ) {
       msg.delete();
     }
